--- apps/journeys-admin/src/components/VisitorInfo/JourneyList/index.ts
@@ -1,1 +1,0 @@
-export { JourneyList } from './JourneyList'
