+++ apps/journeys-admin/src/components/VisitorInfo/VisitorInfo.spec.tsx
@@ -2,7 +2,6 @@ import { MockedProvider } from '@apollo/client/testing'
 import { fireEvent, render, waitFor, within } from '@testing-library/react'
 import { VisitorInfo } from './VisitorInfo'
 import { VisitorInfoProvider } from './VisitorInfoProvider'
-import { getVisitorEventsMock } from './VisitorJourneyList/VisitorJourneyListData'
 
 jest.mock('@mui/material/useMediaQuery', () => ({
   __esModule: true,
@@ -13,7 +12,7 @@ describe('VisitorInfo', () => {
   it.skip('switches between tabs', async () => {
     const { getByText } = render(
       <VisitorInfoProvider>
-        <MockedProvider mocks={[getVisitorEventsMock]}>
+        <MockedProvider mocks={[]}>
           <VisitorInfo id="visitorId" />
         </MockedProvider>
       </VisitorInfoProvider>
