--- apps/journeys-admin/src/components/VisitorInfo/VisitorJourneyList/VisitorJourneyListItem/index.ts
@@ -1,1 +1,0 @@
-export { VisitorJourneyListItem } from './VisitorJourneyListItem'
