--- apps/journeys-admin/src/components/VisitorInfo/VisitorJourneyList/index.ts
@@ -1,1 +1,0 @@
-export { VisitorJourneyList } from './VisitorJourneyList'
