+++ apps/journeys-admin/src/components/VisitorInfo/VisitorJourneysList/EventsCard/CompactEvent/CompactEvent.spec.tsx
