+++ apps/journeys-admin/src/components/VisitorInfo/VisitorJourneysList/EventsCard/CompactEvent/index.ts
