+++ apps/journeys-admin/src/components/VisitorInfo/VisitorJourneysList/EventsCard/GenericEvent/GenericEvent.spec.tsx
