+++ apps/journeys-admin/src/components/VisitorInfo/VisitorJourneysList/EventsCard/GenericEvent/index.ts
