+++ apps/journeys-admin/src/components/VisitorInfo/VisitorJourneysList/index.ts
@@ -1,0 +1,1 @@
+export { VisitorJourneysList } from './VisitorJourneysList'
