+++ apps/journeys-admin/src/components/VisitorInfo/VisitorJourneysList/utils/messagePlatformToLabel/index.ts
