+++ apps/journeys-admin/src/components/VisitorInfo/VisitorJourneysList/utils/messagePlatformToLabel/messagePlatformToLabel.ts
